More assorted review/cleanup/refactor
