+++ test/integration/acceptance/custom/hello_policy/hello_policy_test.go
@@ -497,14 +497,23 @@ func testHelloPolicy(t *testing.T, pub1, pub2 *base.ClusterContext) {
 	// Running the scenarios
 	t.Run("init", func(t *testing.T) { cli.RunScenariosParallel(t, initSteps) })
 	//	mainSteps = mainSteps
-	removeCrd(t, pub1)
+	_, err := removeCrd(pub1)
+	if err != nil {
+		t.Fatalf("CRD removal failed; the test cannot proceed: %v", err)
+	}
 	t.Run("No CRD, all works", func(t *testing.T) { cli.RunScenarios(t, mainSteps) })
 	t.Run("Re-expose service, for next test", func(t *testing.T) { cli.RunScenarios(t, []cli.TestScenario{exposeSteps}) })
 	applyCrd(pub1)
 	// TODO: can this run in parallel?
 	t.Run("CRD added and no policy, all comes down", func(t *testing.T) { cli.RunScenarios(t, checkStuffCameDown) })
 	log.Print("Removing CRD again, some resources should come back up")
-	removeCrd(t, pub1)
+	changed, err := removeCrd(pub1)
+	if err != nil {
+		t.Fatalf("CRD removal failed; the test cannot proceed: %v", err)
+	}
+	if !changed {
+		t.Errorf("CRD removal didn't fail, but also didn't change anything.  Something is probably off")
+	}
 	// TODO: can this run in parallel?
 	t.Run("CRD removed, link should come back up", func(t *testing.T) { cli.RunScenarios(t, checkStuffCameBackUp) })
 	t.Run("closing", func(t *testing.T) { cli.RunScenariosParallel(t, deleteSteps) })
