+++ test/integration/acceptance/custom/hello_policy/main_test.go
@@ -6,14 +6,10 @@ package hello_policy
 import (
 	"log"
 	"os"
-	"regexp"
 	"strings"
 	"sync"
 	"testing"
 
-	"github.com/skupperproject/skupper/api/types"
-	skupperv1 "github.com/skupperproject/skupper/pkg/apis/skupper/v1alpha1"
-	clientv1 "github.com/skupperproject/skupper/pkg/generated/client/clientset/versioned/typed/skupper/v1alpha1"
 	"github.com/skupperproject/skupper/pkg/kube"
 	"github.com/skupperproject/skupper/test/utils/base"
 	"github.com/skupperproject/skupper/test/utils/constants"
@@ -21,222 +17,9 @@ import (
 	"github.com/skupperproject/skupper/test/utils/skupper/cli"
 	"gotest.tools/assert"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 var testPath = "./tmp/"
-
-// Adds the CRD to the cluster
-// TODO: can this be improved?
-func applyCrd(cluster *base.ClusterContext) (err error) {
-	var out []byte
-	log.Printf("Adding CRD into the %v cluster", cluster.KubeConfig)
-	out, err = cluster.KubectlExec("apply -f ../../../../../api/types/crds/skupper_cluster_policy_crd.yaml")
-	if err != nil {
-		log.Printf("CRD applying failed: %v", err)
-		log.Print("Output:\n", string(out))
-		return
-	}
-	return
-}
-
-func isCrdInstalled(cluster *base.ClusterContext) (installed bool, err error) {
-	var out []byte
-	installed = true
-
-	// TODO: replace this by some kube API?
-	out, err = cluster.KubectlExec("get crd skupperclusterpolicies.skupper.io")
-	if err != nil {
-		if strings.Contains(
-			string(out),
-			`Error from server (NotFound): customresourcedefinitions.apiextensions.k8s.io "`) {
-			installed = false
-			err = nil
-		}
-	}
-	return
-}
-
-// Remove the CRD from the cluster
-func removeCrd(t *testing.T, cluster *base.ClusterContext) (changed bool, err error) {
-	changed = true
-
-	log.Printf("Removing CRD from the cluster %v", cluster.KubeConfig)
-
-	installed, err := isCrdInstalled(cluster)
-	if err != nil {
-		t.Fatalf("Failed checking for CRD")
-		return
-	}
-
-	if !installed {
-		changed = false
-		log.Print("CRD was not present, so not changing anything")
-		return
-	}
-
-	if _, err := cluster.KubectlExec("delete crd skupperclusterpolicies.skupper.io"); err != nil {
-		t.Fatalf("Removal of CRD failed: %v", err)
-	}
-	return
-}
-
-// Remove the cluster role, but do not fail if it is not there
-func removeClusterRole(cluster *base.ClusterContext) (changed bool, err error) {
-	changed = true
-	log.Printf("Removing cluster role %v from the CRD definition", types.ControllerServiceAccountName)
-
-	// Is it there?
-	role, err := cluster.VanClient.KubeClient.RbacV1().ClusterRoles().Get(types.ControllerServiceAccountName, metav1.GetOptions{})
-	if role == nil && err != nil {
-		log.Print("The role did not exist on the cluster; skipping removal")
-		changed = false
-		err = nil
-		return
-	}
-	cluster.VanClient.KubeClient.RbacV1().ClusterRoles().Delete(types.ControllerServiceAccountName, nil)
-	return
-}
-
-// Removes all policies from the cluster.
-//
-// If policies are provided, only those will be removed, instead
-func removePolicies(t *testing.T, cluster *base.ClusterContext, policies ...string) (err error) {
-
-	log.Print("Removing policies")
-
-	var list *skupperv1.SkupperClusterPolicyList
-
-	skupperCli, err := clientv1.NewForConfig(cluster.VanClient.RestConfig)
-	if err != nil {
-		return
-	}
-
-	if len(policies) == 0 {
-		policies = []string{}
-		// We're listing and removing everything
-		list, err = listPolicies(t, cluster)
-		if err != nil {
-			return
-		}
-		for _, item := range list.Items {
-			policies = append(policies, item.Name)
-		}
-	}
-
-	for _, item := range policies {
-		log.Printf("- %v", item)
-		item_err := skupperCli.SkupperClusterPolicies().Delete(item, &metav1.DeleteOptions{})
-		if item_err != nil {
-			log.Printf("  removal failed: %v", item_err)
-			if err == nil {
-				// We'll return the first error from the list, but keep trying the others
-				err = item_err
-			}
-		}
-	}
-
-	return
-}
-
-func listPolicies(t *testing.T, cluster *base.ClusterContext) (list *skupperv1.SkupperClusterPolicyList, err error) {
-
-	installed, err := isCrdInstalled(cluster)
-	if err != nil {
-		t.Fatalf("Failed to check for CRD on the cluster")
-		return
-	}
-
-	if !installed {
-		log.Print("The CRD is not installed, so considering the policy list as empty")
-		list = &skupperv1.SkupperClusterPolicyList{}
-		return
-	}
-
-	skupperCli, err := clientv1.NewForConfig(cluster.VanClient.RestConfig)
-	if err != nil {
-		return
-	}
-
-	list, err = skupperCli.SkupperClusterPolicies().List(metav1.ListOptions{})
-	if err != nil {
-		log.Print("Failed listing policies")
-		return
-	}
-
-	return
-}
-
-func keepPolicies(t *testing.T, cluster *base.ClusterContext, patterns []regexp.Regexp) (err error) {
-
-	policyList := []string{}
-
-	list, err := listPolicies(t, cluster)
-	if err != nil {
-		return
-	}
-
-	for _, item := range list.Items {
-		var matched bool
-		for _, re := range patterns {
-			if re.MatchString(item.Name) {
-				matched = true
-				break
-			}
-		}
-		if !matched {
-			policyList = append(policyList, item.Name)
-		}
-	}
-
-	if len(policyList) == 0 {
-		return
-	}
-
-	err = removePolicies(t, cluster, policyList...)
-
-	return
-
-}
-
-// Apply a SkupperClusterPolicySpec with the given name on the
-// requested cluster
-func applyPolicy(name string, spec skupperv1.SkupperClusterPolicySpec, cluster *base.ClusterContext) (err error) {
-
-	log.Printf("Applying policy %v (%v)...", name, spec)
-	skupperCli, err := clientv1.NewForConfig(cluster.VanClient.RestConfig)
-	if err != nil {
-		return
-	}
-	var policy = skupperv1.SkupperClusterPolicy{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "SkupperClusterPolicy",
-			APIVersion: "skupper.io/v1alpha1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: spec,
-	}
-
-	existing, err := skupperCli.SkupperClusterPolicies().Get(name, metav1.GetOptions{})
-	if err != nil {
-		log.Printf("... as a new policy")
-		_, err = skupperCli.SkupperClusterPolicies().Create(&policy)
-		if err != nil {
-			return err
-		}
-	} else {
-		log.Printf("... as an update to an existing policy")
-		policy.ObjectMeta.ResourceVersion = existing.ObjectMeta.ResourceVersion
-		_, err := skupperCli.SkupperClusterPolicies().Update(&policy)
-		if err != nil {
-			return err
-		}
-	}
-
-	return
-}
 
 // Helper to deploy an application on Kubernetes and wait for it
 type KubeDeploy struct {
@@ -246,20 +29,25 @@ type KubeDeploy struct {
 }
 
 // Actual deployment, according to struct's specification, on the
-// given context
+// given context.  If a deployment with the same name already exists,
+// it will just log and keep going.
 func (kd KubeDeploy) deploy(ctx *base.ClusterContext) (err error) {
-	frontend, _ := k8s.NewDeployment(kd.name, ctx.Namespace, k8s.DeploymentOpts{
+	deployment, _ := k8s.NewDeployment(kd.name, ctx.Namespace, k8s.DeploymentOpts{
 		Image:         kd.image,
 		Labels:        kd.labels,
 		RestartPolicy: corev1.RestartPolicyAlways,
 	})
 	// Creating deployments
-	if _, err = ctx.VanClient.KubeClient.AppsV1().Deployments(ctx.Namespace).Create(frontend); err != nil {
+	if _, err = ctx.VanClient.KubeClient.AppsV1().Deployments(ctx.Namespace).Create(deployment); err != nil {
+		if strings.Contains(err.Error(), `deployments.apps "`+kd.name+`" already exists`) {
+			log.Printf("Ignoring application already deployed: %v", err)
+			err = nil
+		}
 		return
 	}
 	// Waiting for deployments to be ready
-	if _, err := kube.WaitDeploymentReady(kd.name, ctx.Namespace, ctx.VanClient.KubeClient, constants.ImagePullingAndResourceCreationTimeout, constants.DefaultTick); err != nil {
-		return err
+	if _, err = kube.WaitDeploymentReady(kd.name, ctx.Namespace, ctx.VanClient.KubeClient, constants.ImagePullingAndResourceCreationTimeout, constants.DefaultTick); err != nil {
+		return
 	}
 	return
 }
@@ -288,12 +76,21 @@ func deployBackend(ctx *base.ClusterContext) (err error) {
 // deployResources Deploys the hello-world-frontend and hello-world-backend
 // pods and validate they are available
 func deployResources(pub *base.ClusterContext, prv *base.ClusterContext) error {
-	deployFrontend(pub)
-	deployBackend(prv)
+	err := deployFrontend(pub)
+	if err != nil {
+		return err
+	}
+	err = deployBackend(prv)
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
 
+// If the prefix is empty, return name, unchanged
+//
+// Else, return "prefix-name"
 func prefixName(prefix, name string) (newName string) {
 
 	newName = name
@@ -373,9 +170,11 @@ func setup(t *testing.T) (pub1, pub2, pub3, prv1 *base.ClusterContext) {
 // the test can be better understood.
 //
 // To run a single test, refer to it as '//testName'.  For example,
-//   go test -run "//testNamespace"
-// This will ensure that not only the tests, but also the setup and tear down
-// are run.
+//
+//   go test -tags policy -timeout 60 -run "//testNamespace"
+//
+// This will ensure that not only the selected tests, but also the setup and
+// tear down are run.
 //
 // TBD control setup/teardown with environment variables.
 //
@@ -413,7 +212,10 @@ func TestPolicies(t *testing.T) {
 					go func() {
 						defer wg.Done()
 						log.Printf("Removing Policy CRD from context %v", context.Namespace)
-						removeCrd(t, context)
+						_, err := removeCrd(context)
+						if err != nil {
+							log.Printf("Failed removing CRD: %v", err)
+						}
 						removeClusterRole(context)
 					}()
 				}
